fix(particle): coerce integer math results
